requirements.txt
annotated-types==0.5.0
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
emoji==2.7.0
flake8==6.1.0
fonttools==4.42.0
fpdf2==2.7.5
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
jellyfish==0.11.2
Jinja2==3.1.2
jsonschema
jsonschema-specifications==2023.7.1
lib50
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pluggy==1.2.0
ptyprocess==0.7.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pyfiglet==0.8.post1
pyflakes==3.1.0
pytest==7.4.0
PyYAML==6.0.1
referencing==0.30.2
requests
rpds-py==0.9.2
soupsieve==2.4.1
tabulate==0.9.0
termcolor==1.1.0
typing_extensions==4.7.1
urllib3==2.0.4
validator-collection==1.5.0